+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+* Prevent `chronometrist-kv-accept` from changing the file if there are no key-values.
+
 ## [0.3.2] - 2019-11-23
 ### Fixed
 * Regression in `chronometrist-value-history-populate`