+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Fixed
-* Prevent `chronometrist-kv-accept` from changing the file if there are no key-values.
+* `chronometrist-kv-accept` will not modify the file if there are no key-values.
+* Regression in `chronometrist-value-history-populate`
 
 ## [0.3.2] - 2019-11-23
 ### Fixed