+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.1] - 2020-01-12
+### Fixed
+* Various declarations for MELPA
+
 ## [0.4.0] - 2019-11-29
 ### Added
 * Custom variable `chronometrist-activity-indicator` to change how an active task is indicated.